pagination: refactored code, made custom useMovies hook
