+++ src/utils/MainApi.js
@@ -65,6 +65,6 @@ class Api {
   }
 }
 
-const api = new Api(API_CONFIG);
+const mainApi = new Api(API_CONFIG);
 
-export default api;
+export default mainApi;
